Fixed minor issues and solved mayor two players mod bugs

+++ FlappyBird_Galli/src/Menu.cpp
@@ -63,9 +63,9 @@ namespace Menu
 		buttonDraw(twoPlayers);
 		buttonDraw(creditsButton);
 
-		DrawText("Play", playButton.Position.x + 20, playButton.Position.y, 20, BLACK);
-		DrawText("Two Players", twoPlayers.Position.x + 20, twoPlayers.Position.y, 20, BLACK);
-		DrawText("Credits", creditsButton.Position.x + 20, creditsButton.Position.y, 20, BLACK);
+		DrawText("Play", static_cast<int>(playButton.Position.x) + 20, static_cast<int>(playButton.Position.y), 20, BLACK);
+		DrawText("Two Players", static_cast<int>(twoPlayers.Position.x) + 20, static_cast<int>(twoPlayers.Position.y), 20, BLACK);
+		DrawText("Credits", static_cast<int>(creditsButton.Position.x) + 20, static_cast<int>(creditsButton.Position.y), 20, BLACK);
 		DrawText("0.3", 5, 5, 10, WHITE);
 		EndDrawing();
