+++ FlappyBird_Galli/src/Obstacle.cpp
@@ -4,6 +4,7 @@ namespace obstacle
 {
     void obstacle::Movement(Obstacle obstacle[2])
     {
+
         obstacle[0].position.x -= 1 * GetFrameTime() * Game::speed;
         obstacle[1].position.x -= 1 * GetFrameTime() * Game::speed;
 
@@ -11,8 +12,8 @@ namespace obstacle
         {
             obstacle[0].position.x = program::screenWidth;
             obstacle[1].position.x = program::screenWidth;
-            obstacle[0].position.y = (rand() % (program::screenHeight - 100)) + 100;
-            obstacle[1].position.y = (obstacle[0].position.y - program::screenHeight) - 150;
+            obstacle[0].position.y = static_cast<float>((rand() % (static_cast<int>(program::screenHeight - 100.0))) + 100.0);
+            obstacle[1].position.y = static_cast<float>((obstacle[0].position.y - program::screenHeight) - 150.0);
         }
     }
 }
